Initial Commit: Basis for the GUI started. Need to add a settings window to add custom sounds/buttons.
Buttons may be added via the .java file -- have an array of hashes or an array of arrays, that contain the sound file directory and name. For every item in the root array, a button is added to the main center of GUI.

Possibly a row of media play buttons at the bottom.